extract BinanceApiClient to manage the request/response between lean and binance
